Create a channel wrapper for bunny/ march hare to cache exchanges
Bunny does this by default, but march hare does not. As far as I can
tell not caching this is really expensive. We are unable to provide a
timeout with this operation and under high rabbitmq load we see very
high publisher latency. My hypothesis is that caching this will allow us
to stick to operations that take a timeout argument.