Check rabbitmq connection status and reconnect when disconnected
The underlying client should be good about reconnecting, but we are
seeing behavior where a runtime (in this case JRuby) is using a lot of
mem and then fails to disconnect despite every other connection
(thousands of others) working just fine. Perhaps it's safer to reconnect
and let the publishers figure it out?

Right now we see:

```
com.rabbitmq.client.AlreadyClosedException: connection is already closed due to connection error; cause: java.io.EOFException
```

This is called when the in-mem publisher tries to create a new channel
(when a consumer thread is created).

TODO: Will this potentially leak a connection? Should we explicitly
disconnect the old connection?