Remove minimum Psych constraint
Rails < 6.0 uses the old way of loading YAML files. This breaks loading
database configs with aliases (a common development practice) with Psych
4.x. Adding a minimum version of Psych to the gemspec here forces an
upgrade to Psych 4.x, which means we end up having to pin Psych to 3.3.2.

Rails 6.1 solves this problem by checking if the new unsafe_load method
exists, and falling back to load if it doesn't. That's a better way.

Resolves #63.